requirements.txt
# 심리 상담 챗봇 + 도서 추천 시스템 의존성

# 데이터 모델
pydantic>=2.5.0

# AI 및 LLM
anthropic>=0.40.0

# CrewAI - Multi-Agent Framework
crewai>=0.28.0
crewai-tools>=0.8.0

# HTTP 요청
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 웹 UI
gradio>=4.0.0
